renamed: api/posts.ts -> app/api/posts/route.ts
멍청실수함
